boj/p15k/p15900/p15904_UCPC는_무엇의_약자일까？_test.go
package p15900

import (
	"bufio"
	"fmt"
	"strings"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestSolve15904(t *testing.T) {
	t.Log("https://www.acmicpc.net/problem/15904")

	//goland:noinspection SpellCheckingInspection
	tests := []struct {
		give string
		want string
	}{
		{
			"Union of Computer Programming Contest club contest",
			"I love UCPC",
		},
		{
			"University Computer Programming",
			"I hate UCPC",
		},
	}

	for i, tt := range tests {
		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
			scanner := bufio.NewScanner(strings.NewReader(tt.give))
			buf := new(strings.Builder)
			writer := bufio.NewWriter(buf)

			Solve15904(scanner, writer)

			err := writer.Flush()
			assert.NoError(t, err)

			got := buf.String()
			assert.Equal(t, tt.want, got)
		})
	}
}
